+++ src/components/settings/RegisterSettings.vue
@@ -2,13 +2,9 @@
   <div class="flex flex-col items-start p-12">
 
     <div class="flex items-center flex-row gap-2 max-w-full w-full mt-2">
-      <div class="tooltip tooltip-right" :data-tip="$t('emailsFileTips')">
-        <span class="font-bold">{{ $t('emailsFile') }}: </span>
-        <font-awesome-icon :icon="['fas', 'circle-exclamation']" />
-      </div>
-      <input type="text" placeholder="example: C:/Users/Administrator/Desktop/emails.txt"
-        class="input input-sm grow input-bordered" v-model="settings.emails_file" />
-      <button class="btn btn-sm btn-info ml-2" @click="selectEmails">{{ $t('select') }}</button>
+      <span class="font-bold">{{ $t('emails') }}: </span>
+      <textarea class="textarea textarea-success grow  h-32 leading-tight" :placeholder="$t('emailsTips')"
+        autocomplete="off" v-model="settings.emails"> </textarea>
     </div>
     <div class="flex items-center flex-row gap-2 max-w-full w-full mt-2">
       <span class="font-bold">{{ $t('password') }}: </span>
@@ -25,7 +21,6 @@
 </template>
 <script>
 import MyButton from '../Button.vue'
-import { open } from '@tauri-apps/api/dialog';
 export default {
   name: 'app',
   components: {
@@ -38,19 +33,6 @@ export default {
     }
   },
   methods: {
-
-    async selectEmails() {
-      const filePath = await open({
-        multiple: false, // 是否允许多选文件
-        directory: false, // 是否选择目录
-        filters: [ // 文件过滤器
-          { name: 'Email File', extensions: ['txt'] },
-        ]
-      });
-      console.log('Selected file path:', filePath);
-      // 将 filePath 用于其他操作
-      this.settings.emails_file = filePath
-    },
     async get_settings() {
       this.$service.get_settings().then(res => {
         this.settings = res.data