+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 3. 修复注册停在选择生日的bug / Fix registration stopped at selecting birthday bug
 4. 移除批量关注功能 / Remove BatchFO function
 5. 修复账号编辑bug / Fix account editing bug
+6. 添加自动重试功能 / Add auto retry function